tools(mlflow): split extractor internals into modules
Keep the single 'extract_permission_events.py --source ...' CLI but move
the implementation out of one 799-line file into focused modules, each
well under the 600-line guideline:

- _permission_common.py  - shared classifiers, follow-up, schema, writers
- _source_mlflow.py       - headless (MLflow) reader + summary
- _source_transcripts.py  - interactive (transcript) reader + summary
- extract_permission_events.py - thin CLI dispatching on --source

The source modules import shared helpers as sibling top-level modules
(resolved at runtime via sys.path[0] when run as a script); the test
loader adds the tools dir to sys.path and merges the modules for access.
